+++ tools/azure-sdk-tools/ci_tools/functions.py
@@ -3,6 +3,7 @@ import subprocess
 import shutil
 import zipfile
 import tarfile
+import stat
 from ast import Not
 from packaging.specifiers import SpecifierSet
 from packaging.version import Version, parse, InvalidVersion
@@ -542,6 +543,28 @@ def find_whl(whl_dir: str, pkg_name: str, pkg_version: str) -> str:
             sys.exit(1)
 
     return None
+
+
+def error_handler_git_access(func, path, exc):
+    """
+    This function exists because the git idx file is written with strange permissions that prevent it from being
+    deleted. Due to this, we need to register an error handler that attempts to fix the file permissions before
+    re-attempting the delete operations.
+    """
+
+    if not os.access(path, os.W_OK):
+        os.chmod(path, stat.S_IWUSR)
+        func(path)
+    else:
+        raise
+
+
+def cleanup_directory(target_directory: str) -> None:
+    """Invokes a directory delete. Specifically handles the case where bad permissions on a git .idx file
+    prevent cleanup of the directory with a generic error.
+    """
+    if os.path.exists(target_directory):
+        shutil.rmtree(target_directory, ignore_errors=False, onerror=error_handler_git_access)
 
 
 def discover_prebuilt_package(dist_directory: str, setup_path: str, package_type: str) -> List[str]: