+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 11.4.0b11 (2023-10-11)
+
+### Features Added
+
+- Added `vector_filter_mode` support for `Search` method.
+- Exposed `VectorizableTextQuery` in `azure.search.document.models`.
+
 ## 11.4.0b10 (2023-10-10)
 
 ### Breaking Changes